Crate bin

Structs

BfsIter
An iterator for a breadth first traversal of a graph.

DfsIter
An iterator for a depth first traversal of a graph.

DiGraph
DiGraph<N, E> is a directed graph, with generic node values N and edge weights E.

Graph
Graph<N, E> is a regular graph, with generic node values N and edge weights E.

MinScored
MinScored<K, T> holds a score K and a scored object T in a pair for use with a BinaryHeap.

Node

NodeCell

Ptr
A reference that is hashed and compared by its pointer value.

Functions

dijkstra
Dijkstra's shortest path algorithm.